requirements.txt
appdirs==1.4.3
asgiref==3.2.10
astroid==2.3.3
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
distlib==0.3.0
Django==3.0.7
django-cors-headers==3.4.0
django-dotenv==1.4.1
djangorestframework==3.11.0
filelock==3.0.12
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
passlib==1.7.2
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
Unipath==1.1
urllib3==1.25.9
virtualenv==20.0.18
virtualenvwrapper-win==1.2.6
wrapt==1.11.2
